Edit design userList and autoReg
