+++ kotlinClient/src/jsMain/kotlin/contentPages/content/solutionsPages/TaskOrUserSolutions.kt
@@ -34,7 +34,7 @@ class TaskOrUserSolutions(
                         title.content = "Решения задачи: ${taskOrUserSolutions.name}"
                     }
                     if (taskOrUserSolutions.solutions.isEmpty()) {
-                        add(Div("Решения не найдены"))
+                        add(Div("Решения не найдены", className = "not-found"))
                     } else {
                         add(AllSolutionsViewer(taskOrUserSolutions.solutions, routing))
                     }
