+++ kotlinClient/src/jsMain/kotlin/contentPages/content/userListPage/UserListViewer.kt
@@ -1,6 +1,7 @@
 package ru.yarsu.contentPages.content.userListPage
 
-import io.kvision.core.onClick
+import io.kvision.html.ButtonStyle
+import io.kvision.html.button
 import io.kvision.html.div
 import io.kvision.panel.VPanel
 import io.kvision.routing.Routing
@@ -12,9 +13,11 @@ class UserListViewer(
 ) : VPanel(className = "UserList") {
     init {
         for (user in userList) {
-            div("${user.surname} ${user.name} (${user.login})")
-            div("Решения пользователя", className = "task-link").onClick {
-                routing.navigate("solution-list/user/${user.id}")
+            div(className = "user-block") {
+                div("${user.surname} ${user.name} (${user.login})")
+                button("Решения пользователя", style = ButtonStyle.LINK).onClick {
+                    routing.navigate("solution-list/user/${user.id}")
+                }
             }
         }
     }
