+++ kotlinClient/src/jsMain/resources/css/UserList.css
@@ -1,7 +1,51 @@
-.UserList .task-link {
-    font-size: 16px;
+.main-page div:has(> .user-block) {
+    display: grid !important;
+    grid-template-columns: repeat(4, 1fr);
+    gap: 2.5rem 2rem;
 }
 
-.actionPanel {
-    margin-right: 20px;
+.user-block {
+    display: flex;
+    flex-direction: column;
+    align-items: stretch;
+    row-gap: 0.75rem;
+    width: 100%;
+}
+
+.user-block > div {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    box-sizing: border-box;
+    min-height: 14rem;
+    padding: 1rem;
+    background-color: #FFFFFF;
+    border: 0.0625rem solid rgba(0, 0, 0, 0.5);
+    border-radius: 0.6875rem;
+    font-size: 1.75rem;
+    font-weight: normal;
+    text-align: center;
+    overflow-wrap: anywhere;
+}
+
+.user-block > .btn {
+    white-space: normal;
+}
+
+@media (max-width: 90rem) {
+    .main-page div:has(> .user-block) {
+        grid-template-columns: repeat(3, 1fr);
+    }
+}
+
+@media (max-width: 64rem) {
+    .main-page div:has(> .user-block) {
+        grid-template-columns: repeat(2, 1fr);
+    }
+}
+
+@media (max-width: 48rem) {
+    .main-page div:has(> .user-block) {
+        grid-template-columns: 1fr;
+    }
 }
